step2

+++ src/main/scala/data/ItemRepository.scala
@@ -4,6 +4,6 @@ package data
   * A real repo/mock could extend this trait
   */
 trait ItemRepository {
-  def getItemCost(itemName : String) : Int
+  def getItemCost(itemName : String) : BigDecimal
 }
